restructure cli directory
