+++ cli/texttools/__init__.py
@@ -1,3 +1,3 @@
-from . import Compiler, Validator
+from . import cli, compiler, validator
 
-__all__ = ["Compiler", "Validator"]
+__all__ = ["cli", "compiler", "validator"]
